requirements.txt
# pydantic-ai - AI Agent框架
# 注意: mcp和openai extra有冲突,先安装mcp版本,再单独安装openai
pydantic-ai-slim[mcp]==1.34.0
openai>=1.0.0
httpx>=0.20.0

flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
ebooklib>=0.17.1
beautifulsoup4>=4.12.0
python-multipart>=0.0.6
mobi>=0.4.1
PyMuPDF>=1.23.0
fastmcp==2.13.0
Pillow>=10.0.0
asgiref>=3.8.0
reportlab>=4.0.0
